+++ src/Reactioncraft/Net/Common/ItemPieceNet.java
@@ -18,14 +18,16 @@ public class ItemPieceNet extends Item
 	}
 
 	@Override
-	public String getItemDisplayName(ItemStack par1ItemStack) {
+	public String getItemDisplayName(ItemStack par1ItemStack) 
+	{
 		return "Net Piece";
 	}
 
 	@Override
-	public void addInformation(ItemStack itemStack, EntityPlayer player, List list, boolean par4) {
+	public void addInformation(ItemStack itemStack, EntityPlayer player, List list, boolean par4) 
+	{
 		if (itemStack.stackTagCompound != null)
-			list.add("Level: " + itemStack.stackTagCompound.getByte("str"));
+			list.add("Level: " + itemStack.stackTagCompound.getInteger("str"));
 		else
 			list.add("Please craft to see results");
 		super.addInformation(itemStack, player, list, par4);
